Trait arduino_mega2560::prelude::_

pub trait _ {
    type Error;
    pub fn set_low(&mut self) -> Result<(), Self::Error>;

    pub fn set_high(&mut self) -> Result<(), Self::Error>;
}

Single digital push-pull output pin

Associated Types

type Error

Error type

Loading content...
Required methods

pub fn set_low(&mut self) -> Result<(), Self::Error>

Drives the pin low

NOTE the actual electrical state of the pin may not actually be low, e.g. due to external electrical sources

pub fn set_high(&mut self) -> Result<(), Self::Error>

Drives the pin high

NOTE the actual electrical state of the pin may not actually be high, e.g. due to external electrical sources

Loading content...